requirements.txt
# Employee Onboarding Guide Builder — Python 3.11+
streamlit>=1.29.0
google-genai>=1.0.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
openai>=1.10.0
jsonschema>=4.19.0
